requirements.txt
alembic==1.11.1
appdirs==1.4.4
atomicwrites==1.4.0
cloudinary==1.44.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.100.0
jsonpointer==2.1
lxml==5.2.1
matplotlib==3.9.2
mkl-service==2.4.0
passlib==1.7.4
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1-modules==0.2.8
pydantic==2.0.3
pydantic-settings==2.0.1
pydantic_core==2.3.0
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pywin32==305.1
rsa==4.9.1
setuptools==75.1.0
SQLAlchemy==2.0.19
starlette==0.27.0
uvicorn==0.22.0
wheel==0.44.0
